pyproject.toml
[tool.poetry]
name = "warehouse"
version = "0.1.0"
description = "warehouse web app"
authors = ["mbrav <mbrav@protonmail.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.8"
wagtail = "^2.15"
django-dotenv = "^1.4.2"
django-measurement = "^3.2.3"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
